remove multiple directory iteration

+++ LogAnalyzerLibrary.API/Controllers/ArchiveController.cs
@@ -13,11 +13,17 @@ namespace LogAnalyzerLibrary.API.Controllers
         [HttpPost]
         public async Task<IActionResult> ArchiveLogs([FromBody] PeriodDTO request)
         {
-            if (request.DirectoryPaths == null || !request.DirectoryPaths.Any())
+            if (request.DirectoryPath == null || !request.DirectoryPath.Any())
             {
                 // Return a 400 Bad Request if the directory paths are empty or null
                 return BadRequest("Directory paths cannot be null or empty.");
             }
+
+            if (request.StartDate > request.EndDate)
+            {
+                return BadRequest("Start date cannot be later than the end date.");
+            }
+
 
             try
             {
@@ -30,14 +36,15 @@ namespace LogAnalyzerLibrary.API.Controllers
             }
         }
 
-        [HttpDelete("{startDate}/{endDate}")]
-        public async Task<IActionResult> DeleteArchives([FromRoute] DateTime startDate, [FromRoute] DateTime endDate, [FromBody] PeriodDTO model)
+        [HttpDelete]
+        public async Task<IActionResult> DeleteArchives([FromQuery] PeriodDTO model)
         {
             try
             {
-
-                model.StartDate = startDate;
-                model.EndDate = endDate;
+                if (model.StartDate > model.EndDate)
+                {
+                    return BadRequest("Start date cannot be later than the end date.");
+                }
 
                 string result = await archiveService.DeleteArchiveAsync(model);
                 return Ok(result);
