+++ LogAnalyzerLibrary.Application/ArchiveService/ArchiveService.cs
@@ -2,91 +2,78 @@
 using LogAnalyzerLibrary.Application.Shared;
 using System.IO.Compression;
 
-namespace LogAnalyzerLibrary.Application.ArchiveService;
-
-public class ArchiveService : IArchiveService
+namespace LogAnalyzerLibrary.Application.ArchiveService
 {
-    public async Task<List<string>> ArchiveLogsAsync(PeriodDTO model)
+    public class ArchiveService : IArchiveService
     {
-        // List to store paths of created zip files
-        List<string> createdZipFiles = [];
-
-        // Iterate over each directory path provided
-        foreach (var directoryPath in model.DirectoryPaths)
+        public async Task<List<string>> ArchiveLogsAsync(PeriodDTO model)
         {
+            List<string> createdZipFiles = [];
+
+            var directoryPath = model.DirectoryPath;
             if (Directory.Exists(directoryPath))
             {
-                // Get log files within the date range from the current directory
                 var logFiles = Directory.GetFiles(directoryPath, "*.log")
                     .Where(file => FileInDateRange.IsFileInDateRange(file, model.StartDate, model.EndDate))
                     .ToList();
 
                 if (logFiles.Count > 0)
                 {
-                    // Define the zip file name based on the date range and directory
                     string zipFileName = $"{model.StartDate:dd_MM_yyyy}-{model.EndDate:dd_MM_yyyy}.zip";
                     string zipFilePath = Path.Combine(directoryPath, zipFileName);
 
-                    // Create the zip file
                     await Task.Run(() =>
                     {
                         using var zipArchive = ZipFile.Open(zipFilePath, ZipArchiveMode.Create);
                         foreach (var logFile in logFiles)
                         {
                             zipArchive.CreateEntryFromFile(logFile, Path.GetFileName(logFile));
-                            File.Delete(logFile); // Delete the original log file after adding to the zip
+                            File.Delete(logFile);
                         }
                     });
 
                     createdZipFiles.Add(zipFilePath);
                 }
             }
-        }
 
-        if (createdZipFiles.Count == 0)
-        {
-            return ["No logs found for the specified date range."];
-        }
-
-        return createdZipFiles;
-    }
-
-
-    public async Task<string> DeleteArchiveAsync(PeriodDTO model)
-    {
-        bool anyFilesDeleted = false;
-        List<string> deletedFiles = [];
-
-        foreach (var directoryPath in model.DirectoryPaths)
-        {
-            await Task.Run(() =>
+            if (createdZipFiles.Count == 0)
             {
-                // Get all zip files in the directory and its subdirectories
-                var zipFiles = Directory.GetFiles(directoryPath, "*.zip", SearchOption.AllDirectories);
+                return new List<string> { "No logs found for the specified date range." };
+            }
 
-                foreach (var zipFile in zipFiles)
-                {
-                    // Check if the file creation date is within the specified date range
-                    if (FileInDateRange.IsFileInDateRange(zipFile, model.StartDate, model.EndDate))
-                    {
-                        // Delete the file
-                        File.Delete(zipFile);
-                        anyFilesDeleted = true;
-
-                        // Add the name of the deleted file to the list
-                        deletedFiles.Add(Path.GetFileName(zipFile));
-                    }
-                }
-            });
+            return createdZipFiles;
         }
 
-        if (!anyFilesDeleted)
+        public async Task<string> DeleteArchiveAsync(PeriodDTO model)
         {
-            throw new FileNotFoundException("No zip files found in the specified date range.");
+            bool anyFilesDeleted = false;
+            List<string> deletedFiles = new List<string>();
+
+            var directoryPath = model.DirectoryPath;
+            if (Directory.Exists(directoryPath))
+            {
+                await Task.Run(() =>
+                {
+                    var zipFiles = Directory.GetFiles(directoryPath, "*.zip", SearchOption.AllDirectories);
+
+                    foreach (var zipFile in zipFiles)
+                    {
+                        if (FileInDateRange.IsFileInDateRange(zipFile, model.StartDate, model.EndDate))
+                        {
+                            File.Delete(zipFile);
+                            anyFilesDeleted = true;
+                            deletedFiles.Add(Path.GetFileName(zipFile));
+                        }
+                    }
+                });
+            }
+
+            if (!anyFilesDeleted)
+            {
+                throw new FileNotFoundException("No zip files found in the specified date range.");
+            }
+
+            return $"Archive deleted successfully. Deleted files: {string.Join(", ", deletedFiles)}";
         }
-
-        // Return the names of the deleted archives
-        return $"Archive deleted successfully. Deleted files: {string.Join(", ", deletedFiles)}";
-
     }
 }
